text added and the big clean
