--- crates/bottomless-pit/src/camera.rs
@@ -1,223 +1,0 @@
-use winit::event::{WindowEvent, KeyboardInput, ElementState, VirtualKeyCode};
-use wgpu::util::DeviceExt;
-
-pub struct Camera {
-    eye: cgmath::Point3<f32>,
-    target: cgmath::Point3<f32>,
-    up: cgmath::Vector3<f32>,
-    aspect: f32,
-    fovy: f32,
-    znear: f32,
-    zfar: f32,
-    pub(crate) uniform: CameraUniform,
-    pub(crate) buffer: wgpu::Buffer,
-    pub(crate) bind_group: wgpu::BindGroup,
-}
-
-#[rustfmt::skip]
-pub const OPENGL_TO_WGPU_MATRIX: cgmath::Matrix4<f32> = cgmath::Matrix4::new(
-    1.0, 0.0, 0.0, 0.0,
-    0.0, 1.0, 0.0, 0.0,
-    0.0, 0.0, 0.5, 0.0,
-    0.0, 0.0, 0.5, 1.0,
-);
-
-impl Camera {
-    pub fn new(eye: (f32, f32, f32), target: (f32, f32, f32), up: cgmath::Vector3<f32>, aspect_ratio: f32, fovy: f32, znear: f32, zfar: f32, device: &wgpu::Device) -> Self {
-        let uniform = CameraUniform::new();
-
-        let buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-            label: Some("Camera Buffer"),
-            contents: bytemuck::cast_slice(&[uniform]),
-            usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
-        });
-        
-        let bind_group_layout = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-            entries: &[
-                wgpu::BindGroupLayoutEntry {
-                    binding: 0,
-                    visibility: wgpu::ShaderStages::VERTEX,
-                    ty: wgpu::BindingType::Buffer {
-                        ty: wgpu::BufferBindingType::Uniform,
-                        has_dynamic_offset: false,
-                        min_binding_size: None,
-                    },
-                    count: None,
-                }
-            ],
-            label: Some("camera_bind_group_layout"),
-        });
-
-        let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-            layout: &bind_group_layout,
-            entries: &[
-                wgpu::BindGroupEntry {
-                    binding: 0,
-                    resource: buffer.as_entire_binding(),
-                }
-            ],
-            label: Some("camera_bind_group"),
-        });
-
-        Self {
-            eye: eye.into(),
-            target: target.into(),
-            up,
-            aspect: aspect_ratio,
-            fovy,
-            znear,
-            zfar,
-            uniform,
-            buffer,
-            bind_group,
-        }
-    }
-
-    pub fn build_projection_matrix(&self) -> cgmath::Matrix4<f32> {
-        cgmath::perspective(cgmath::Deg(self.fovy), self.aspect, self.znear, self.zfar)
-    }
-
-    pub fn build_view_projection_matrix(&self) -> cgmath::Matrix4<f32> {
-        // this moves the world to be at the position and roation of the camera
-        // an inverse of whatever the transform matrix of the camera would be
-        let view = cgmath::Matrix4::look_at_rh(self.eye, self.target, self.up);
-        // this warps it so it looks like depth is real.
-        let projection = self.build_projection_matrix();
-        // converts cgmaths OpenGL values to wgpus values
-        OPENGL_TO_WGPU_MATRIX * projection * view
-    }
-
-    pub fn resize(&mut self, new_size: winit::dpi::PhysicalSize<u32>) {
-        self.aspect = new_size.width as f32 / new_size.height as f32;
-    }
-
-    pub fn update(&mut self, queue: &wgpu::Queue) {
-        self.uniform.update_view_proj(self.build_view_projection_matrix());
-        queue.write_buffer(&self.buffer, 0, bytemuck::cast_slice(&[self.uniform]));
-    }
-
-    pub fn make_bind_group_layout(device: &wgpu::Device) -> wgpu::BindGroupLayout {
-        device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-            entries: &[
-                wgpu::BindGroupLayoutEntry {
-                    binding: 0,
-                    visibility: wgpu::ShaderStages::VERTEX,
-                    ty: wgpu::BindingType::Buffer {
-                        ty: wgpu::BufferBindingType::Uniform,
-                        has_dynamic_offset: false,
-                        min_binding_size: None,
-                    },
-                    count: None,
-                }
-            ],
-            label: Some("camera_bind_group_layout"),
-        })
-    }
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy, bytemuck::Pod, bytemuck::Zeroable)]
-pub struct CameraUniform {
-    view_proj: [[f32; 4]; 4],
-}
-
-impl CameraUniform {
-    pub fn new() -> Self {
-        use cgmath::SquareMatrix;
-        Self {
-            view_proj: cgmath::Matrix4::identity().into(),
-        }
-    }
-
-    pub fn update_view_proj(&mut self, new_matrix: cgmath::Matrix4<f32>) {
-        self.view_proj = new_matrix.into();
-    }
-}
-
-impl Into<CameraUniform> for cgmath::Matrix4<f32> {
-    fn into(self) -> CameraUniform {
-        CameraUniform {
-            view_proj: self.into()
-        }
-    }
-}
-
-pub struct CameraController {
-    speed: f32,
-    is_forward_pressed: bool,
-    is_backward_pressed: bool,
-    is_right_pressed: bool,
-    is_left_pressed: bool,
-}
-
-impl CameraController {
-    pub fn new(speed: f32) -> Self {
-        Self {
-            speed,
-            is_backward_pressed: false,
-            is_forward_pressed: false,
-            is_left_pressed: false,
-            is_right_pressed: false,
-        }
-    }
-
-    pub fn process_events(&mut self, event: &WindowEvent) -> bool {
-        match event {
-            WindowEvent::KeyboardInput {
-                input: KeyboardInput {
-                    state,
-                    virtual_keycode: Some(keycode),
-                    ..
-                },
-                ..
-            } => {
-                let is_pressed = *state == ElementState::Pressed;
-                match keycode {
-                    VirtualKeyCode::W | VirtualKeyCode::Up => {
-                        self.is_forward_pressed = is_pressed;
-                        true
-                    },
-                    VirtualKeyCode::A | VirtualKeyCode::Left => {
-                        self.is_left_pressed = is_pressed;
-                        true
-                    },
-                    VirtualKeyCode::S | VirtualKeyCode::Back => {
-                        self.is_backward_pressed = is_pressed;
-                        true
-                    },
-                    VirtualKeyCode::D | VirtualKeyCode::Right => {
-                        self.is_right_pressed = is_pressed;
-                        true
-                    },
-                    _ => false,
-                }
-            },
-            _ => false,
-        }
-    }
-
-    pub fn update_camera(&self, camera: &mut Camera) {
-        use cgmath::InnerSpace;
-        let forward = camera.target - camera.eye;
-        let forward_norm = forward.normalize();
-        let forward_mag = forward.magnitude();
-
-        if self.is_forward_pressed && forward_mag > self.speed {
-            camera.eye += forward_norm * self.speed;
-        }
-        if self.is_backward_pressed {
-            camera.eye -= forward_norm * self.speed;
-        }
-
-        let right = forward_norm.cross(camera.up);
-        let forward = camera.target - camera.eye;
-        let forward_mag = forward.magnitude();
-
-        if self.is_right_pressed {
-            camera.eye = camera.target - (forward + right * self.speed).normalize() * forward_mag;
-        }
-        if self.is_left_pressed {
-            camera.eye = camera.target - (forward - right * self.speed).normalize() * forward_mag;
-        }
-    }
-}
