+++ crates/bottomless-pit/src/rect.rs
@@ -1,4 +1,5 @@
-use crate::{Vertex, cache::TextureIndex};
+use crate::Vertex;
+use crate::vertex::vert_pixels_to_screenspace;
 use crate::vectors::Vec2;
 
 pub const RECT_INDICIES: &[u16] = &[
@@ -25,6 +26,20 @@ impl Rectangle {
         }
     }
 
+    pub fn from_pixels(pos: Vec2<f32>, size: [f32; 2], colour: [f32; 4], screen_size: Vec2<u32>) -> Self {
+        let pos = pos.to_raw();
+        let points = [
+            vert_pixels_to_screenspace(Vertex::from_2d(pos, [0.0, 0.0], colour), screen_size), 
+            vert_pixels_to_screenspace(Vertex::from_2d([pos[0] + size[0], pos[1]], [1.0, 0.0], colour), screen_size),
+            vert_pixels_to_screenspace(Vertex::from_2d([pos[0] + size[0], pos[1] + size[1]], [1.0, 1.0], colour), screen_size),
+            vert_pixels_to_screenspace(Vertex::from_2d([pos[0], pos[1] + size[1]], [0.0, 1.0], colour), screen_size),
+        ];
+
+        Self { 
+            points
+        }
+    }
+
     pub fn from_points(points: [[f32; 2]; 4], colour: [f32; 4]) -> Self {
         let points = [
             Vertex::from_2d(points[0], [0.0, 0.0], colour),
@@ -44,28 +59,6 @@ impl Rectangle {
         let pos = Vec2{x: corners[0][0], y: corners[0][1]};
         
         Self::new(pos, [width, height], colour)
-    }
-
-    pub fn get_width(&self) -> f32 {
-        let left = self.points[0].get_pos_2d()[0];
-        let right = self.points[1].get_pos_2d()[0];
-
-        if left < 0.0 && right > 0.0 {
-            left.abs() + right
-        } else {
-            (right - left).abs()
-        }
-    }
-
-    pub fn get_height(&self) -> f32 {
-        let top = self.points[0].get_pos_2d()[1];
-        let bottom = self.points[2].get_pos_2d()[1];
-
-        if top < 0.0 && bottom > 0.0 {
-            top.abs() + bottom
-        } else {
-            (bottom - top).abs()
-        }
     }
 
     pub fn get_vertices(&self) -> [Vertex; 4] {
