+++ crates/bottomless-pit/src/vertex.rs
@@ -1,3 +1,5 @@
+use crate::Vec2;
+
 #[repr(C)]
 #[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
 pub struct Vertex {
@@ -38,14 +40,6 @@ impl Vertex {
             colour,
         }
     }
-
-    pub fn get_pos(&self) -> [f32; 3] {
-        self.position
-    }
-
-    pub fn get_pos_2d(&self) -> [f32; 2] {
-        [self.position[0], self.position[1]]
-    }
 }
 
 #[repr(C)]
@@ -82,4 +76,21 @@ impl LineVertex {
             colour,
         }
     }
+}
+
+pub(crate) fn vert_pixels_to_screenspace(mut point: Vertex, screen_size: Vec2<u32>) -> Vertex {
+    //println!("{:?}", point);
+    let width = screen_size.x as f32;
+    let height = screen_size.y as f32;
+    point.position[0] = (2.0 * point.position[0] / width) - 1.0;
+    point.position[1] = ((2.0 * point.position[1] / height) - 1.0) * -1.0;
+    point
+}
+
+pub(crate) fn line_vert_pixels_to_screenspace(mut point: LineVertex, screen_size: Vec2<u32>) -> LineVertex {
+    let width = screen_size.x as f32;
+    let height = screen_size.y as f32;
+    point.pos[0] = (2.0 * point.pos[0] / width) - 1.0;
+    point.pos[1] = ((2.0 * point.pos[1] / height) - 1.0) * -1.0;
+    point
 }
